requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
requests
beautifulsoup4
selenium
flask
dash
psycopg2-binary